refactor(classes):put methods to specialized classes
